reorganised the project Changes to be committed: 	modified:   main.py 	new file:   src/extras.py 	new file:   src/model_build.py 	modified:   src/utils.py
